feat(bkm): eşiklere istatistiksel sınama + literatür kıyası; raf kaybı ölçütü değişti
Kullanıcı kabul görmüş yöntem ve literatür istedi. Sonuç: bir ölçüt reddedildi,
biri değişti, iki yapısal sınır ortaya çıktı.

ÖLÇÜT SINAMASI (Cochran-Armitage trend testi + Wilson skor aralığı, geriye dönük
as-of 01.08.2025, sonuç = sonraki 12 ayda satıldı mı):
· merkez stoğu ekseni: chi2(1)=95,57 p=1,43e-22 AMA monoton değil
  (3,2/4,9/8,2/7,4/6,4/11,9) ve iki komşu bant GA'sı çakışık → REDDEDİLDİ
· önceki yıl talebi: chi2(1)=873,01 p=7,24e-192, monoton
  (3,5/13,3/38,4/57,4/60,0), 1-4 [11,6-15,2] ile 5-19 [32,4-44,9] ayrık → SEÇİLDİ
Raf Bulunurluk Kaybı ölçütü MerkezStok>=5 yerine SatisToplam>=5; kart 3.464 →
448 çeşit / 2.069.830 ₺. Talep kanıtı olmayan 8.736 çeşit (%3,5 satma oranı)
bulunurluk kaybı değil ölü stok.
DERS: küçük p bir ölçütü doğrulamaz — CA yalnız doğrusal trende karşı güçlü.

VERİDEN SEÇİLEN KESİMİN BEDELİ beyan edildi (Altman & Royston 2006 BMJ;
Royston/Altman/Sauerbrei 2006 Stat Med): kesim farkı abartır, spuriously
significant üretir. Kesimdeki oran farkı artık etki ölçüsü olarak sunulmuyor.

YENİ BULGU — aralıklı talep: 12 tam ayda satış olan ay sayısına göre hiç satmayan
%44,1 · 1-2 ay %27,9 · düzgün talep yalnız %8,1 (ADI<=1,33, Syntetos/Boylan/
Croston eşiği). "Günlük ortalama satış" ve "gün-stok" çeşitlerin %91,9'unda
yanıltıcı; Croston/SBA/TSB ayrıştırması yapılmadı, beyan edildi.

SEKTÖR KIYASI: ölçülen yıllık devir Kitap 1,28 · Kırtasiye 1,25 · Dergi 5,95;
yayınlanmış kitapçı hedefi 3,0-4,0. Kıyas ÇIKARIM olarak etiketli (ticari kaynak,
BKM saf kitapçı değil, GMROI hesabı standart değil). Eşik sektöre ÇEKİLMEDİ —
gerekçesi panelde yazılı.

İKİ SINIR: kayıt doğruluğu ölçülmüş değil (panel %0,12 yalnız aritmetik imkânsız
kayıtlar; literatürde IRI %50-70, DeHoratius & Raman %65) · kayıp satış tahminleri
alt sınır (sağdan sansürlü talep; EM düzeltmesi uygulanmadı).

Kalıcılaştırma: .claude/rules/olctum-mu-cikardim-mi.md § EŞİK TÜRETME (4 kural) ·
sema'ya 4 kayıt · arşiv sorgular/2026-09-10-esik-istatistik-ve-talep-deseni.sql.

Doğrulama: build 0 hata · sema_degismez 39/39 yeşil.

+++ dashboard/Models/SatisAnaliziModels.cs
@@ -249,7 +249,7 @@ public sealed record SatisAnaliziKpi(
     /// </summary>
     int RafBosCesit,
     decimal RafBosTutar,
-    int RafBosSatisliCesit,
+    long RafBosSatisliCesit,
     int VeriKirliCesit,
     decimal VeriKirliTutar,
     /// <summary>YENİ ÜRÜN — hareketsiz/aşırı ölçütlerinin KASITLI dışladığı kova.
